+++ src/TinyGsmClientSIM7000SSL.h
@@ -868,7 +868,7 @@ class TinyGsmSim7000SSL
   // Between the modemBeginSend and modemEndSend, modemSend calls:
   // stream.write(reinterpret_cast<const uint8_t*>(buff), len);
   // stream.flush();
-  int16_t modemEndSendImpl(uint16_t len, uint8_t mux) {
+  size_t modemEndSendImpl(size_t len, uint8_t mux) {
     // after posting data, module responds with:
     //+CASEND: <cid>,<result>,<sendlen>
     if (waitResponse(GF(AT_NL "+CASEND:")) != 1) { return 0; }
